+++ src/sc/embedder/tests/test_validator.py
@@ -1,37 +1,26 @@
 # -*- coding: utf-8 -*-
-from sc.embedder.content.embedder import validate_int_or_percentage
-from zope.interface import Invalid
+from sc.embedder.utils import validate_int_or_percentage
 
 import unittest
 
 
 class ValidatorTestCase(unittest.TestCase):
 
-    def test_validate_int_or_percentage(self):
-        # Integer
-        self.assertEqual(validate_int_or_percentage('420'), None)
+    def test_validate_int(self):
+        self.assertTrue(validate_int_or_percentage('420'))
+        # invalid values
+        self.assertFalse(validate_int_or_percentage('-420'))
+        self.assertFalse(validate_int_or_percentage('0'))
+        self.assertFalse(validate_int_or_percentage('1.23'))
+        self.assertFalse(validate_int_or_percentage('123!'))
+        self.assertFalse(validate_int_or_percentage('12%3'))
+        self.assertFalse(validate_int_or_percentage('12 3'))
+        self.assertFalse(validate_int_or_percentage('99999'))
 
-        # Percentage
-        self.assertEqual(validate_int_or_percentage('100%'), None)
-
-        # Invalid values
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('-420')
-
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('123 %')  # spaces not allowed
-
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('1.23')  # not an integer
-
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('123!')
-
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('12%3')
-
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('12 3')
-
-        with self.assertRaises(Invalid):
-            validate_int_or_percentage('99999')  # too big
+    def test_validate_percentage(self):
+        self.assertTrue(validate_int_or_percentage('100%'))
+        # invalid values
+        self.assertFalse(validate_int_or_percentage('-100%'))
+        self.assertFalse(validate_int_or_percentage('0%'))
+        self.assertFalse(validate_int_or_percentage('123 %'))
+        self.assertFalse(validate_int_or_percentage('999%'))